AMB-1895 rename services to lib
