+++ e2e/lib/authentication.py
