+++ e2e/lib/aws.py
